refactor: user 관련 auth 클래스는 user 폴더로 이동
