--- src/main/java/kr/tennispark/auth/application/dto/TokenDTO.java
@@ -1,10 +1,0 @@
-package kr.tennispark.auth.application.dto;
-
-import lombok.Builder;
-
-@Builder
-public record TokenDTO(
-        String refreshToken,
-        String accessToken
-) {
-}
