--- src/main/java/kr/tennispark/auth/application/exception/PhoneNotVerifiedException.java
@@ -1,16 +1,0 @@
-package kr.tennispark.auth.application.exception;
-
-import kr.tennispark.common.utils.ApiUtils;
-import org.springframework.http.HttpStatus;
-
-public class PhoneNotVerifiedException extends RuntimeException {
-    private final String message = "아직 휴대폰 인증이 완료되지 않았습니다.";
-
-    public ApiUtils.ApiResult<?> body() {
-        return ApiUtils.error(HttpStatus.BAD_REQUEST, message);
-    }
-
-    public HttpStatus status() {
-        return HttpStatus.BAD_REQUEST;
-    }
-}
