--- src/main/java/kr/tennispark/auth/infrastructure/exception/SmsSendFailedException.java
@@ -1,16 +1,0 @@
-package kr.tennispark.auth.infrastructure.exception;
-
-import kr.tennispark.common.utils.ApiUtils;
-import org.springframework.http.HttpStatus;
-
-public class SmsSendFailedException extends RuntimeException {
-    private final String message = "메세지 전송에 실패하였습니다.";
-
-    public ApiUtils.ApiResult<?> body() {
-        return ApiUtils.error(HttpStatus.BAD_REQUEST, message);
-    }
-
-    public HttpStatus status() {
-        return HttpStatus.BAD_REQUEST;
-    }
-}
