--- src/main/java/kr/tennispark/auth/infrastructure/sms/SmsProperties.java
@@ -1,12 +1,0 @@
-package kr.tennispark.auth.infrastructure.sms;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "sms")
-public record SmsProperties(
-        String apiKey,
-        String apiSecret,
-        String apiUrl,
-        String fromNumber
-) {
-}
