--- src/main/java/kr/tennispark/auth/presentation/dto/response/RegisterMemberResponse.java
@@ -1,7 +1,0 @@
-package kr.tennispark.auth.presentation.dto.response;
-
-public record RegisterMemberResponse(
-        String accessToken,
-        String refreshToken
-) {
-}
