--- src/main/java/kr/tennispark/auth/presentation/dto/response/ReissueTokenResponse.java
@@ -1,7 +1,0 @@
-package kr.tennispark.auth.presentation.dto.response;
-
-public record ReissueTokenResponse(
-        String accessToken,
-        String refreshToken
-) {
-}
