+++ src/main/java/kr/tennispark/auth/user/application/dto/TokenDTO.java
@@ -1,0 +1,10 @@
+package kr.tennispark.auth.user.application.dto;
+
+import lombok.Builder;
+
+@Builder
+public record TokenDTO(
+        String refreshToken,
+        String accessToken
+) {
+}
