+++ src/main/java/kr/tennispark/auth/user/presentation/dto/response/RegisterMemberResponse.java
@@ -1,0 +1,7 @@
+package kr.tennispark.auth.user.presentation.dto.response;
+
+public record RegisterMemberResponse(
+        String accessToken,
+        String refreshToken
+) {
+}
