+++ src/main/java/kr/tennispark/auth/user/presentation/dto/response/ReissueTokenResponse.java
@@ -1,0 +1,7 @@
+package kr.tennispark.auth.user.presentation.dto.response;
+
+public record ReissueTokenResponse(
+        String accessToken,
+        String refreshToken
+) {
+}
