+++ src/main/java/kr/tennispark/common/exception/GlobalExceptionHandler.java
@@ -3,7 +3,7 @@ package kr.tennispark.common.exception;
 import com.fasterxml.jackson.databind.JsonMappingException.Reference;
 import com.fasterxml.jackson.databind.exc.InvalidFormatException;
 import java.util.stream.Collectors;
-import kr.tennispark.auth.application.exception.ExpiredTokenException;
+import kr.tennispark.auth.user.application.exception.ExpiredTokenException;
 import kr.tennispark.common.exception.base.DuplicateException;
 import kr.tennispark.common.exception.base.InvalidException;
 import kr.tennispark.common.exception.base.NotFoundException;
